+++ demo/admin/src/news/NewsPage.tsx
@@ -11,7 +11,12 @@ import {
     ToolbarAutomaticTitleItem,
     ToolbarBackButton,
 } from "@comet/admin";
+<<<<<<< HEAD
 import { ContentScopeIndicator, useContentScope } from "@comet/cms-admin";
+=======
+import { ContentScopeIndicator } from "@comet/cms-admin";
+import { useContentScope, useContentScopeConfig } from "@src/common/ContentScopeProvider";
+>>>>>>> main
 import { useIntl } from "react-intl";
 
 import { NewsForm } from "./generated/NewsForm";
@@ -31,6 +36,9 @@ const FormToolbar = () => (
 export function NewsPage() {
     const intl = useIntl();
     const { scope } = useContentScope();
+
+    useContentScopeConfig({ redirectPathAfterChange: "/structured-content/news" });
+
     return (
         <Stack topLevelTitle={intl.formatMessage({ id: "news.news", defaultMessage: "News" })}>
             <StackSwitch>