+++ packages/admin/cms-admin/src/dam/blocks/DamFileDownloadLinkBlock.tsx
@@ -1,4 +1,5 @@
 import { gql } from "@apollo/client";
+<<<<<<< HEAD
 import { Field, FinalFormSelect, messages } from "@comet/admin";
 import { Delete } from "@comet/admin-icons";
 import { Box, Divider, MenuItem, Typography } from "@mui/material";
@@ -12,6 +13,16 @@ import { BlocksFinalForm } from "../../blocks/form/BlocksFinalForm";
 import { createBlockSkeleton } from "../../blocks/helpers/createBlockSkeleton";
 import { BlockCategory, type BlockDependency, type BlockInterface } from "../../blocks/types";
 import { DamPathLazy } from "../../form/file/DamPathLazy";
+=======
+import { Field, FinalFormSelect } from "@comet/admin";
+import { AdminComponentPaper, BlockCategory, BlockDependency, BlockInterface, BlocksFinalForm, createBlockSkeleton } from "@comet/blocks-admin";
+import { Divider, MenuItem } from "@mui/material";
+import { deepClone } from "@mui/x-data-grid/utils/utils";
+import { FormattedMessage } from "react-intl";
+
+import { DamFileDownloadLinkBlockData, DamFileDownloadLinkBlockInput } from "../../blocks.generated";
+import { CmsBlockContext } from "../../blocks/CmsBlockContextProvider";
+>>>>>>> main
 import { FileField } from "../../form/file/FileField";
 import { type GQLDamFileDownloadLinkFileQuery, type GQLDamFileDownloadLinkFileQueryVariables } from "./DamFileDownloadLinkBlock.generated";
 
@@ -118,6 +129,7 @@ export const DamFileDownloadLinkBlock: BlockInterface<DamFileDownloadLinkBlockDa
                     openFileType: state.openFileType ?? "Download",
                 }}
             >
+<<<<<<< HEAD
                 {state.file === undefined ? (
                     <Field name="file" component={FileField} fullWidth />
                 ) : (
@@ -134,6 +146,9 @@ export const DamFileDownloadLinkBlock: BlockInterface<DamFileDownloadLinkBlockDa
                         </BlockAdminComponentButton>
                     </BlockAdminComponentPaper>
                 )}
+=======
+                <Field name="file" component={FileField} fullWidth />
+>>>>>>> main
                 <Divider />
                 <BlockAdminComponentPaper>
                     <Field