+++ packages/admin/cms-admin/src/pages/pageTree/PageTree.tsx
@@ -19,8 +19,13 @@ import AutoSizer, { type Size } from "react-virtualized-auto-sizer";
 import { type Align, FixedSizeList as List } from "react-window";
 import { useDebouncedCallback } from "use-debounce";
 
+<<<<<<< HEAD
 import { useContentScope } from "../../contentScope/Provider";
 import { type GQLPagesQuery, type GQLPagesQueryVariables } from "../pagesPage/createPagesQuery";
+=======
+import { usePageTreeScope } from "../config/usePageTreeScope";
+import { GQLPagesQuery, GQLPagesQueryVariables } from "../pagesPage/createPagesQuery";
+>>>>>>> main
 import {
     type GQLMovePageTreeNodesByPosMutation,
     type GQLPagesCacheQuery,
@@ -133,7 +138,7 @@ const PageTree: ForwardRefRenderFunction<PageTreeRefApi, PageTreeProps> = (
     }, [pages]);
 
     const pageTreeService = useMemo(() => new PageTreeService(levelOffsetPx, pages), [pages]);
-    const { scope } = useContentScope();
+    const scope = usePageTreeScope();
     const snackbarApi = useSnackbarApi();
 
     const debouncedSetHoverState = useDebouncedCallback(