+++ packages/admin/cms-admin/src/userPermissions/UserGrid.tsx
@@ -27,6 +27,7 @@ import {
     type GQLUserGridQueryVariables,
 } from "./UserGrid.generated";
 
+<<<<<<< HEAD
 interface UserPermissionsUserGridToolbarProps extends GridToolbarProps {
     toolbarAction: ReactNode;
 }
@@ -40,6 +41,23 @@ function UserPermissionsUserGridToolbar({ toolbarAction }: UserPermissionsUserGr
         </DataGridToolbar>
     );
 }
+=======
+function UserPermissionsGridToolbar({ toolbarAction }: { toolbarAction?: React.ReactNode }) {
+    return (
+        <DataGridToolbar>
+            <ToolbarItem>
+                <GridToolbarQuickFilter />
+            </ToolbarItem>
+            <ToolbarItem>
+                <GridFilterButton />
+            </ToolbarItem>
+            <FillSpace />
+            {toolbarAction && <ToolbarActions>{toolbarAction}</ToolbarActions>}
+        </DataGridToolbar>
+    );
+}
+
+>>>>>>> main
 type Props = {
     toolbarAction?: ReactNode;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -226,8 +244,13 @@ export const UserPermissionsUserGrid = ({ toolbarAction, rowAction, actionsColum
             columns={columns}
             rowCount={data?.users.totalCount ?? 0}
             loading={loading}
+<<<<<<< HEAD
             slots={{
                 toolbar: UserPermissionsUserGridToolbar as GridSlotsComponent["toolbar"],
+=======
+            components={{
+                Toolbar: UserPermissionsGridToolbar,
+>>>>>>> main
             }}
             slotProps={{
                 toolbar: {