+++ packages/api/cms-api/src/file-uploads/file-uploads.service.ts
@@ -5,11 +5,19 @@ import { createHmac } from "crypto";
 import { addHours } from "date-fns";
 import hasha from "hasha";
 import { basename, extname, parse } from "path";
+import { Readable } from "stream";
 
 import { BlobStorageBackendService } from "../blob-storage/backends/blob-storage-backend.service";
+<<<<<<< HEAD
 import { FileUploadInput } from "../file-utils/file-upload.input";
 import { slugifyFilename } from "../file-utils/files.utils";
 import { ALL_TYPES } from "../file-utils/images.constants";
+=======
+import { createHashedPath } from "../blob-storage/utils/create-hashed-path.util";
+import { FileUploadInput } from "../dam/files/dto/file-upload.input";
+import { slugifyFilename } from "../dam/files/files.utils";
+import { ALL_TYPES } from "../dam/images/images.constants";
+>>>>>>> main
 import { DownloadParams, ImageParams } from "./dto/file-uploads-download.params";
 import { FileUpload } from "./entities/file-upload.entity";
 import { FileUploadsConfig } from "./file-uploads.config";
@@ -89,5 +97,23 @@ export class FileUploadsService {
         const filename = parse(file.name).name;
 
         return ["/file-uploads", hash, file.id, timeout, resizeWidth, filename].join("/");
+    }
+
+    async getFileContent(file: FileUpload): Promise<Buffer> {
+        const filePath = createHashedPath(file.contentHash);
+        const fileExists = await this.blobStorageBackendService.fileExists(this.config.directory, filePath);
+
+        if (!fileExists) {
+            throw new Error("File not found");
+        }
+
+        const stream = await this.blobStorageBackendService.getFile(this.config.directory, filePath);
+
+        const chunks: Buffer[] = [];
+        for await (const chunk of stream as Readable) {
+            chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
+        }
+        const buffer = Buffer.concat(chunks);
+        return buffer;
     }
 }