+++ packages/admin/admin-date-time/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/admin-date-time
 
+<<<<<<< HEAD
 ## 8.0.0-beta.5
 
 ### Patch Changes
@@ -101,6 +102,40 @@
 - Updated dependencies [682a674]
     - @comet/admin@8.0.0-beta.0
     - @comet/admin-icons@8.0.0-beta.0
+=======
+## 7.25.0
+
+### Patch Changes
+
+-   @comet/admin@7.25.0
+-   @comet/admin-icons@7.25.0
+
+## 7.24.0
+
+### Patch Changes
+
+-   fa611381b: Don't show the `ClearInputAdornment` if the `TimePicker` is disabled
+-   fa611381b: Support the `disabled` prop in the `DateTimePicker`
+-   Updated dependencies [fc900f217]
+    -   @comet/admin@7.24.0
+    -   @comet/admin-icons@7.24.0
+
+## 7.23.0
+
+### Patch Changes
+
+-   @comet/admin@7.23.0
+-   @comet/admin-icons@7.23.0
+
+## 7.22.0
+
+### Patch Changes
+
+-   Updated dependencies [2cf573b72]
+-   Updated dependencies [086774f01]
+    -   @comet/admin@7.22.0
+    -   @comet/admin-icons@7.22.0
+>>>>>>> main
 
 ## 7.21.1