+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/eslint-config
 
+<<<<<<< HEAD
 ## 8.0.0-beta.5
 
 ### Major Changes
@@ -66,6 +67,31 @@
 ### Minor Changes
 
 - a8edddb: Enable `@typescript-eslint/consistent-type-imports` rule
+=======
+## 7.25.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.25.0
+
+## 7.24.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.24.0
+
+## 7.23.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.23.0
+
+## 7.22.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.22.0
+>>>>>>> main
 
 ## 7.21.1