+++ packages/site/site-nextjs/CHANGELOG.md
@@ -1,10 +1,43 @@
 # @comet/site-nextjs
 
+<<<<<<< HEAD
 ## 8.0.0-beta.5
 
 ### Patch Changes
 
 - @comet/site-react@8.0.0-beta.5
+=======
+## 7.25.0
+
+### Minor Changes
+
+-   b421ed273: Support captions in the `DamVideoBlock`
+
+    The captions can be set uploaded as .vtt files and linked to videos in the DAM.
+
+### Patch Changes
+
+-   @comet/site-react@7.25.0
+
+## 7.24.0
+
+### Patch Changes
+
+-   @comet/site-react@7.24.0
+
+## 7.23.0
+
+### Patch Changes
+
+-   @comet/site-react@7.23.0
+
+## 7.22.0
+
+### Patch Changes
+
+-   Updated dependencies [e4327e250]
+    -   @comet/site-react@7.22.0
+>>>>>>> main
 
 ## 7.21.1